requirements.txt
discord.py>=2.1.0
dotenv>=0.21.1